# swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"title": "Diabetes 360 API",
		"version": "1.0.0",
		"description": "Comprehensive API documentation for Diabetes 360 health management application",
		"contact": {
			"name": "API Support"
		}
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Enter JWT token"
			}
		},
		"schemas": {
			"Error": {
				"type": "object",
				"properties": {
					"status": {
						"type": "number",
						"example": 400,
						"description": "HTTP status code"
					},
					"success": {
						"type": "boolean",
						"example": false
					},
					"message": {
						"type": "string",
						"example": "Error message",
						"description": "Error message description"
					},
					"data": {
						"description": "Error response data - null for most errors, empty array [] for array endpoints",
						"oneOf": [
							{
								"type": "null"
							},
							{
								"type": "array",
								"items": {}
							},
							{
								"type": "object"
							}
						],
						"example": null
					}
				},
				"required": ["status", "success", "message"]
			},
			"SuccessResponse": {
				"type": "object",
				"properties": {
					"status": {
						"type": "number",
						"example": 200,
						"description": "HTTP status code"
					},
					"success": {
						"type": "boolean",
						"example": true
					},
					"data": {
						"description": "Response data payload"
					},
					"message": {
						"type": "string",
						"example": "Operation successful",
						"description": "Success message"
					}
				},
				"required": ["status", "success", "message"]
			},
			"User": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"example": "uuid-string"
					},
					"firstName": {
						"type": "string",
						"example": "John"
					},
					"lastName": {
						"type": "string",
						"example": "Doe"
					},
					"email": {
						"type": "string",
						"example": "john@example.com"
					},
					"emailVerified": {
						"type": "boolean",
						"example": false
					},
					"provider": {
						"type": "string",
						"example": "manual"
					},
					"providerId": {
						"type": "string",
						"nullable": true
					},
					"avatar": {
						"type": "string",
						"nullable": true
					},
					"role": {
						"type": "string",
						"enum": ["customer", "admin", "physician"],
						"example": "customer"
					},
					"paymentType": {
						"type": "string",
						"enum": ["free", "monthly", "annual"],
						"example": "free"
					},
					"isActive": {
						"type": "boolean",
						"example": true
					},
					"profileComplete": {
						"type": "boolean",
						"example": false,
						"description": "Indicates if the user profile is complete (for customer role)"
					},
					"profileData": {
						"oneOf": [
							{
								"$ref": "#/components/schemas/CustomerData"
							},
							{
								"$ref": "#/components/schemas/PhysicianData"
							}
						],
						"nullable": true,
						"description": "User profile data - CustomerData for customers, PhysicianData for physicians, null if profile not complete"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"CustomerData": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"example": "uuid-string"
					},
					"userId": {
						"type": "string",
						"example": "uuid-string"
					},
					"gender": {
						"type": "string",
						"enum": ["male", "female"],
						"example": "male"
					},
					"birthday": {
						"type": "string",
						"format": "date-time",
						"example": "1990-01-15T00:00:00Z"
					},
					"diagnosisDate": {
						"type": "string",
						"format": "date-time",
						"example": "2020-05-10T00:00:00Z"
					},
					"weight": {
						"type": "string",
						"example": "70",
						"description": "Weight in kg as string"
					},
					"height": {
						"type": "string",
						"example": "175",
						"description": "Height in cm as string"
					},
					"diabetesType": {
						"type": "string",
						"enum": ["type1", "type2", "gestational", "prediabetes"],
						"example": "type2"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
			"PhysicianData": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string",
						"example": "uuid-string"
					},
					"userId": {
						"type": "string",
						"example": "uuid-string"
					},
					"specialtyId": {
						"type": "string",
						"example": "uuid-string"
					},
					"specialty": {
						"type": "string",
						"example": "Endocrinology",
						"description": "Specialty name (from physician_specialties table)",
						"nullable": true
					},
					"practiceStartDate": {
						"type": "string",
						"format": "date-time",
						"example": "2010-06-01T00:00:00Z"
					},
					"consultationFee": {
						"type": "string",
						"example": "1500.00",
						"description": "Consultation fee as string"
					},
					"imageUrl": {
						"type": "string",
						"nullable": true,
						"example": "https://example.com/image.jpg"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				}
			},
...
		}
	},
...
	"paths": {
...
		"/api/user/profile": {
			"get": {
				"summary": "Get current user's profile with profile data",
				"tags": ["User"],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"responses": {
					"200": {
						"description": "Profile retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/SuccessResponse"
										},
										{
											"type": "object",
											"properties": {
												"data": {
													"$ref": "#/components/schemas/User"
												}
											},
											"example": {
												"status": 200,
												"success": true,
												"message": "Profile retrieved successfully",
												"data": {
													"id": "uuid-string",
													"firstName": "John",
													"lastName": "Doe",
													"email": "john@example.com",
													"emailVerified": false,
													"provider": "manual",
													"providerId": null,
													"avatar": null,
													"role": "customer",
													"tier": "free",
													"isActive": true,
													"profileComplete": true,
													"profileData": {
														"id": "uuid-string",
														"userId": "uuid-string",
														"gender": "male",
														"birthday": "1990-01-15T00:00:00Z",
														"diagnosisDate": "2020-05-10T00:00:00Z",
														"weight": "70",
														"height": "175",
														"diabetesType": "type2",
														"createdAt": "2024-01-01T00:00:00Z",
														"updatedAt": "2024-01-01T00:00:00Z"
													},
													"createdAt": "2024-01-01T00:00:00Z",
													"updatedAt": "2024-01-01T00:00:00Z"
												}
											}
										}
									]
								},
								"examples": {
									"customer": {
										"summary": "Customer user",
										"value": {
											"status": 200,
											"success": true,
											"message": "Profile retrieved successfully",
											"data": {
												"id": "uuid-string",
												"firstName": "John",
												"lastName": "Doe",
												"email": "john@example.com",
												"emailVerified": false,
												"provider": "manual",
												"providerId": null,
												"avatar": null,
												"role": "customer",
												"tier": "free",
												"isActive": true,
												"profileComplete": true,
												"profileData": {
													"id": "uuid-string",
													"userId": "uuid-string",
													"gender": "male",
													"birthday": "1990-01-15T00:00:00Z",
													"diagnosisDate": "2020-05-10T00:00:00Z",
													"weight": "70",
													"height": "175",
													"diabetesType": "type2",
													"createdAt": "2024-01-01T00:00:00Z",
													"updatedAt": "2024-01-01T00:00:00Z"
												},
												"createdAt": "2024-01-01T00:00:00Z",
												"updatedAt": "2024-01-01T00:00:00Z"
											}
										}
									},
									"physician": {
										"summary": "Physician user",
										"value": {
											"status": 200,
											"success": true,
											"message": "Profile retrieved successfully",
											"data": {
												"id": "uuid-string",
												"firstName": "Alice",
												"lastName": "Smith",
												"email": "alice@example.com",
												"emailVerified": true,
												"provider": "manual",
												"providerId": null,
												"avatar": null,
												"role": "physician",
												"tier": "paid",
												"isActive": true,
												"profileComplete": true,
												"profileData": {
													"id": "uuid-string",
													"userId": "uuid-string",
													"specialtyId": "uuid-string",
													"specialty": "Endocrinology",
													"practiceStartDate": "2010-05-01T00:00:00Z",
													"consultationFee": "100",
													"imageUrl": "https://example.com/physician.jpg",
													"createdAt": "2024-01-01T00:00:00Z",
													"updatedAt": "2024-01-01T00:00:00Z"
												},
												"createdAt": "2024-01-01T00:00:00Z",
												"updatedAt": "2024-01-01T00:00:00Z"
											}
										}
									},
									"incomplete": {
										"summary": "Profile incomplete (profileData is null)",
										"value": {
											"status": 200,
											"success": true,
											"message": "Profile retrieved successfully",
											"data": {
												"id": "uuid-string",
												"firstName": "Joe",
												"lastName": "Bloggs",
												"email": "joe@example.com",
												"emailVerified": false,
												"provider": "manual",
												"providerId": null,
												"avatar": null,
												"role": "customer",
												"tier": "free",
												"isActive": true,
												"profileComplete": false,
												"profileData": null,
												"createdAt": "2024-01-01T00:00:00Z",
												"updatedAt": "2024-01-01T00:00:00Z"
											}
										}
									}
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized - invalid or missing token",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					},
					"403": {
						"description": "Forbidden - insufficient permissions",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Error"
								}
							}
						}
					}
				}
			}
		}
	}
}